now check if cat is only non zero weight in questions

+++ src/views/surveys.py
@@ -469,6 +469,17 @@ def delete_category():
 
     category_id = request.form["category_id"]
     survey_id = request.form["survey_id"]
+    cat_name = request.form["cat_name"]
+    questions = survey_service.get_questions_of_survey(survey_id)
+    category_required_by = helper.questions_where_given_category_is_required(questions, cat_name)
+
+    if len(category_required_by) != 0:
+        message = f"Check the following question's category weights before deleting '{cat_name}': '"
+        message += "', '".join(category_required_by)
+        message += "'"
+        flash(message, "error")
+        return redirect(f"/surveys/{survey_id}")
+
     return_value = survey_service.delete_category(category_id)
     if return_value is True:
         flash("Succesfully deleted category", "confirmation")
